:art: Restructure code. Mostly copied from @MurdeRM3LODY
