+++ src/decorations/borders.rs
@@ -1,7 +1,4 @@
-use crate::{
-	libs::structs::decorations::BorderShader,
-	CONFIG,
-};
+use crate::CONFIG;
 use smithay::{
 	backend::renderer::{
 		element::Element,
@@ -33,6 +30,11 @@ const ROUNDED_BORDER_FRAG: &str = include_str!("shaders/rounded_corners.frag");
 const BORDER_FRAG: &str = include_str!("shaders/borders.frag");
 
 struct BorderShaderElements(RefCell<HashMap<Window, PixelShaderElement>>);
+
+pub struct BorderShader {
+	pub rounded: GlesPixelProgram,
+	pub default: GlesPixelProgram,
+}
 
 impl BorderShader {
 	pub fn init(renderer: &mut GlowRenderer) {
