+++ src/decorations/shaders/rounded_corners.frag
