--- src/libs/mod.rs
@@ -1,9 +1,0 @@
-pub mod backends;
-pub mod config;
-pub mod decorations;
-pub mod handlers;
-pub mod layouts;
-pub mod state;
-pub mod structs;
-pub mod tiling;
-pub mod workspaces;
