--- src/libs/structs/args.rs
@@ -1,8 +1,0 @@
-use clap::Parser;
-
-#[derive(Parser, Debug)]
-#[command(author, version, about, long_about = None)]
-pub struct Args {
-	#[arg(short, long)]
-	pub backend: String,
-}
