--- src/libs/structs/decorations.rs
@@ -1,6 +1,0 @@
-use smithay::backend::renderer::gles::GlesPixelProgram;
-
-pub struct BorderShader {
-	pub rounded: GlesPixelProgram,
-	pub default: GlesPixelProgram,
-}
