--- src/libs/structs/mod.rs
@@ -1,5 +1,0 @@
-pub mod args;
-pub mod comms;
-pub mod decorations;
-pub mod state;
-pub mod workspaces;
